Implement GitLab Pages
1. convert README.markdown -> README.org
2. reuse README.org in org-mode based GitLab pages implementation

This way, we use org-mode to maintain README which GitLab-CI turns into
a homepage on GitLab Pages. It doesn't require a separate git branch and
makes it effortless to keep README and landing page in sync.  We don't
check in any generated/exported files, it's all done by GitLab-CI during
publishing.