Avoid timeout in non_empty_list_shrink_test
On Gitlab's CI sometimes the test fails with a timeout. When run locally it
takes about 2 seconds to run on a relatively new laptop. So on a heavly loaded
CI environment it's plausible it would take more than 5 seconds sometime.

Reduce the number of tries from 25 to 15. Now it takes under a second to run,
hopefully that gives it enough headroom in the CI environment.